code/nautilus_v2/nautilus/requirements.txt
dacite>=1,<2
docker>=7.1,<7.2
jaydebeapi>=1.2,<1.3
omegaconf>=2.2,<2.4
pandas>=2.2,<2.3
pympler>=1.0,<1.1
pyyaml>=6.0,<6.1
ray[default]==2.9.3
xmltodict>=0.13,<0.14
requests==2.31.0
redis>=5.0,<5.1